% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{filter_swiss_protein}
\alias{filter_swiss_protein}
\title{filter_swiss_protein}
\usage{
filter_swiss_protein(swiss_protein, probability = 0.1)
}
\arguments{
\item{swiss_protein}{swiss_protein tibble.}

\item{probability}{Probability threshold.}
}
\value{
A new swiss_protein tibble.
}
\description{
Filter swiss_protein based on probability
}
\examples{
swiss_protein <- filter_swiss_protein(setsList$swiss_protein, probability = 0.9)
}
