% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{grsa}
\alias{grsa}
\title{GRSA}
\usage{
grsa(input, enrich_tibble, adjust = "fdr", thread = 1, perm = 499)
}
\arguments{
\item{input}{Please edit the input as example_data.}

\item{enrich_tibble}{Please edit the enrich_tibble as setsList's function set.}

\item{adjust}{Method of P-value correction.See stats::p.adjust.}

\item{thread}{The number of parallel threads.}

\item{perm}{The number of resampling times when calculating a random distribution.}
}
\value{
An enrichmentRes tibble.
}
\description{
The function set enrichment P-value is calculated using GRSA algorithm.
}
\examples{
data("example_data", package = "MetaboRich")
data("metabolitesList", package = "MetaboRich")
data("setsList", package = "MetaboRich")
example_data <- id_mapping(example_data, from = "kegg_id", "hmdb_id", metabolites_tibble = metabolitesList$hmdb)
enrichmentRes <- grsa(input = example_data, enrich_tibble = setsList$smpdb_pathway, adjust = "fdr", thread = 8)
}
