% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{id_mapping}
\alias{id_mapping}
\title{id_mapping}
\usage{
id_mapping(input, from = "kegg_id", to = "hmdb_id", metabolites_tibble)
}
\arguments{
\item{input}{Please edit the input as example_data.}

\item{from}{id from.}

\item{to}{id to.}

\item{metabolites_tibble}{See metabolitesList.}
}
\value{
An example_data format tibble.
}
\description{
To convert ids between databases,
you need to select appropriate metabolites_tibble based on requirements,
see metabolitesList.
}
\examples{
data("example_data", package = "MetaboRich")
data("metabolitesList", package = "MetaboRich")
example_data <- id_mapping(input = example_data, from = "kegg_id", to = "hmdb_id", metabolites_tibble = metabolitesList$hmdb)
}
