% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{kstest}
\alias{kstest}
\title{KSTESt}
\usage{
kstest(input, enrich_tibble, adjust = "fdr", thread = 1)
}
\arguments{
\item{input}{Please edit the input as example_data.}

\item{enrich_tibble}{Please edit the enrich_tibble as setsList's function set.}

\item{adjust}{Method of P-value correction.See stats::p.adjust.}

\item{thread}{The number of parallel threads.}
}
\value{
An enrichmentRes tibble.
}
\description{
The P-value of the function set of metabolites was calculated using kstest
}
\examples{
data("example_data", package = "MetaboRich")
data("metabolitesList", package = "MetaboRich")
data("setsList", package = "MetaboRich")
example_data <- id_mapping(example_data, from = "kegg_id", "hmdb_id", metabolites_tibble = metabolitesList$hmdb)
enrichmentRes <- kstest(input = example_data, enrich_tibble = setsList$smpdb_pathway, adjust = "fdr", thread = 8)
}
