% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ora}
\alias{ora}
\title{ORA}
\usage{
ora(
  input,
  enrich_tibble,
  enrich_type = "all",
  N_type = "measure",
  adjust = "fdr",
  thread = 1
)
}
\arguments{
\item{input}{Please edit the input as example_data.}

\item{enrich_tibble}{Please edit the enrich_tibble as setsList's function set.}

\item{enrich_type}{"all": Enrichment analysis was performed using all differential metabolites;
"up": Enrichment analysis was performed using upregulated differential metabolites;
"down": Enrichment analysis was performed using downregulated differential metabolites.}

\item{N_type}{"measure": The measured metabolites were used as background;
"database": Use the input metabolite function set as background.}

\item{adjust}{Method of P-value correction.See stats::p.adjust.}

\item{thread}{The number of parallel threads.}
}
\value{
An enrichmentRes tibble.
}
\description{
The classic ORA algorithm uses
hypergeometric distribution for
pvalue calculation of function set.
}
\examples{
data("metabolitesList");data("setsList");data("example_data")
ora(input = example_data, enrich_tibble = setsList$kegg_pathway, enrich_type = "all", N_type = "database", adjust = "fdr", thread = 2)
}
