% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_compareRes}
\alias{plot_compareRes}
\title{plot_compareRes}
\usage{
plot_compareRes(enrichmentResList, top = 15, plot_type = 1)
}
\arguments{
\item{enrichmentResList}{An enrichment result tibble.}

\item{top}{Select the top feature set before the p-value to display.}

\item{plot_type}{Display diagram types, 1,2.}
}
\value{
visNetwork or igraph
}
\description{
The results of various enrichment algorithms are compared by network diagram.
}
\examples{
data("metabolitesList");data("setsList");data("example_data")
enrichmentRes_ora <- ora(input = example_data, enrich_tibble = setsList$kegg_pathway, enrich_type = "all", N_type = "database", adjust = "fdr", thread = 2)
enrichmentRes_kstest <- kstest(input = example_data, enrich_tibble = setsList$kegg_pathway, adjust = "fdr", thread = 2)
enrichmentRes_grsa <- grsa(input = example_data, enrich_tibble = setsList$kegg_pathway, adjust = "fdr", thread = 2)
plot_compareRes(enrichmentResList = list(ora = enrichmentRes_ora, kstest = enrichmentRes_kstest, grsa = enrichmentRes_grsa), plot_type = 1)
}
