% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_enrichmentRes}
\alias{plot_enrichmentRes}
\title{plot_enrichmentRes}
\usage{
plot_enrichmentRes(
  enrichmentRes,
  top = 10,
  plot_type = 1,
  legend.position = "none"
)
}
\arguments{
\item{enrichmentRes}{An enrichment result tibble.}

\item{top}{Select the top feature set before the p-value to display.}

\item{plot_type}{Display diagram types, 1,2,3.}

\item{legend.position}{The position of the diagram column}
}
\value{
A ggplot object.
}
\description{
The results of functional set enrichment analysis were shown.
}
\examples{
data("metabolitesList");data("setsList");data("example_data")
enrichmentRes_ora <- ora(input = example_data, enrich_tibble = setsList$kegg_pathway, enrich_type = "all", N_type = "database", adjust = "fdr", thread = 2)
plot_enrichmentRes(enrichmentRes = enrichmentRes_ora, plot_type = 1)
plot_enrichmentRes(enrichmentRes = enrichmentRes_ora, plot_type = 2)
plot_enrichmentRes(enrichmentRes = enrichmentRes_ora, plot_type = 3)
}
