% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{setsList_protein2setsList_GO}
\alias{setsList_protein2setsList_GO}
\title{setsList_protein2setsList_GO}
\usage{
setsList_protein2setsList_GO(setsList_protein)
}
\arguments{
\item{setsList_protein}{A setsList's protein-related metabolite set.}
}
\value{
A GO metabolite set
}
\description{
Convert a setsList's protein-related metabolite set to a GO metabolite set.
}
\examples{
swiss_protein <- filter_swiss_protein(setsList$swiss_protein, probability = 0.9)
swiss_GO <- setsList_protein2setsList_GO(swiss_protein)
hmdb_GO <- setsList_protein2setsList_GO(setsList_protein = setsList$hmdb_protein)
}
