% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{UNIPROT2GO}
\alias{UNIPROT2GO}
\title{UNIPROT2GO}
\usage{
UNIPROT2GO(uniprot_id)
}
\arguments{
\item{uniprot_id}{A character of uniprot.}
}
\value{
A tibble.
}
\description{
Enter a vector of uniprot and return these uniprot-related GO terms.
}
\examples{
swiss_protein <- filter_swiss_protein(setsList$swiss_protein, probability = 0.9)
uniprot_id <- swiss_protein$id
GO_tibble <- UNIPROT2GO(uniprot_id)
}
